\name{ge.settings}
\alias{ge.settings}
\title{parse settings file}
\usage{
  ge.settings(input.file = NULL)
}
\arguments{
  \item{input.file}{settings file containing information
  needed for GE processing}
}
\description{
  Read settings file for GE processing
}
\examples{
\dontrun{
settings <- settings()
settings <- settings('/home/$USER/settings.xml')
}
}
\author{
  Shawn P. Serbin
}

