% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.raw}
\alias{plot.raw}
\title{A function plot raw A-Ci, A-Q or other analysis curves.  Shows flagged observations.}
\usage{
\method{plot}{raw}(file.dir = NULL, data.file = NULL, out.dir = NULL,
  data = NULL, type = c("A-Ci", "A-Q"))
}
\arguments{
\item{file.dir}{}

\item{data.file}{Input data file to process}

\item{out.dir}{Main output directory for diagnostic plots}

\item{data}{use data already in the environment}

\item{type}{The type of gas exchange data to plot.  Current options: A-Ci and A-Q}

\item{output.type}{.png or .pdf}
}
\description{
A function plot raw A-Ci, A-Q or other analysis curves.  Shows flagged observations.
}
\author{
Shawn P. Serbin
}

