% Generated by roxygen2 (4.0.2): do not edit by hand
\name{arrhenius.scaling}
\alias{arrhenius.scaling}
\title{An Arrhenius temperature scaling function for temperature dependent parameters}
\usage{
arrhenius.scaling(Tleaf.1 = NULL, Tleaf.2 = NULL, Ea = NULL,
  Obs.val = NULL)
}
\arguments{
\item{Tleaf.1}{the temperature (in degrees C) of the observation at the base temperature}

\item{Tleaf.2}{the desired temperature (in degrees C) to scale the observation}

\item{Ea}{activation energy of the Arrhenius equation}

\item{Obs.val}{the value of the observation at the base temperature (i.e. Tleaf.1)}
}
\description{
Simple Arrhenius temperature scaling function
}
\examples{
arrhenius.scaling(Tleaf.1=25,Tleaf.2=30,Ea=45.23,Obs.val=140.4)

# Bernacchi et al.,
Kc <- arrhenius.scaling(Tleaf.1=25,Tleaf.2=30,Ea=79.430,Obs.val=404.9)
Ko <- arrhenius.scaling(Tleaf.1=25,Tleaf.2=30,Ea=36.380,Obs.val=278.4)
Ko <- arrhenius.scaling(Tleaf.1=25,Tleaf.2=30,Ea=36.380,Obs.val=278.4)
Gamma.Star <- arrhenius.scaling(Tleaf.1=25,Tleaf.2=30,Ea=37.830,Obs.val=42.75)
}
\author{
Shawn P. Serbin
}

