\name{read.ge.data}
\alias{read.ge.data}
\title{A function to read in formatted LiCor 6400 data files}
\usage{
  read.ge.data(data.dir = NULL, out.dir = NULL,
    ge.file.ext = ".csv", QC = TRUE,
    output.file.ext = ".csv", ge.dataframe = FALSE,
    settings.file = NULL)
}
\description{
  A function to read in formatted LiCor 6400 data for
  processing
}

