% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.ge.data}
\alias{read.ge.data}
\title{A function to read in formatted LiCor 6400 data files.  Does some inital error
checking including removal of missing value flags (e.g. -99,-999,-9999,-9999.0) and
checking of proper data columns}
\usage{
read.ge.data(file.dir = NULL, data.file = NULL, QC = FALSE)
}
\arguments{
\item{file.dir}{}

\item{data.file}{Optional}

\item{QC}{Run QC checks on data import?  Dafault is QC=FALSE.  This is typically run
as part of the fitting functions.}
}
\description{
A function to read in formatted LiCor 6400 data for processing
}
\examples{
\dontrun{
file.dir <- '/Users/sserbin/Data/GitHub/R-GasExchange/inst/extdata/barrow-aci.csv'
file.dir <- '/Users/sserbin/Data/GitHub/R-GasExchange/inst/extdata/barrow-aci-test.csv'
file.dir <- system.file("extdata/barrow-aci.csv",package="GasExchange")
ge.data <- read.ge.data(file.dir=file.dir)

}
}
\author{
Shawn P. Serbin
}

